fixed scaling and hit, die, restart game in shootout

+++ api/public/shootout/js/Game.js
@@ -7,35 +7,20 @@ Shooter.Game.prototype = {
       this.game.time.advancedTiming = true;
     },
   create: function() {
-    this.map = this.game.add.tilemap('level3');
+    this.map = this.game.add.tilemap('scifi');
+    this.game.scale.scaleMode = Phaser.ScaleManager.SHOW_ALL;
 
-    //the first parameter is the tileset name as specified in Tiled, the second is the key to the asset
-    this.map.addTilesetImage('tiles_spritesheet', 'gameTiles');
+    this.map.addTilesetImage('scifi_spritesheet', 'gameTiles');
 
     //create layers
     this.backgroundlayer = this.map.createLayer('background');
     this.blockedLayer = this.map.createLayer('blocked');
-
-    //collision on blockedLayer
+    this.blockedLayer.resizeWorld();
     this.map.setCollisionBetween(1, 5000, true, 'blocked');
 
-    // this.backgroundlayer.scale.set(0.5);
-    //resizes the game world to match the layer dimensions
-    // this.backgroundlayer.wrap = true;
-    this.backgroundlayer.resizeWorld();
-    // //create coins
-    // this.createCoins();
-
-    //create player
-    // this.player = this.game.add.sprite(100, 300, 'player');
-
-    //enable physics on the player
-    // this.game.physics.arcade.enable(this.player);
-
-    // //player gravity
-    // this.player.body.gravity.y = 1000;
     this.bulletTime = 0;
-    this.lives = 3;
+    this.HITS = 3;
+    this.LIVES = 3;
 
 
     this.bullets = this.game.add.group();
@@ -45,9 +30,6 @@ Shooter.Game.prototype = {
     this.bullets.setAll('anchor.x', 0.5);
     this.bullets.setAll('anchor.y', 1);
 
-    var playerDuckImg = this.game.cache.getImage('playerDuck');
-    var duckedDimensions = {width: playerDuckImg.width, height: playerDuckImg.height};
-
     this.players = [];
     this.spawnPositions = [{x: 100, y: 300},{x: 300, y: 300}, {x: 200, y: 300}]
 
@@ -55,23 +37,24 @@ Shooter.Game.prototype = {
 
     //sounds
     // this.coinSound = this.game.add.audio('coin');
+    this.coinSound = this.game.add.audio('coin');
+    this.hurtSound = this.game.add.audio('hurt');
+    this.shotSound = this.game.add.audio('shot');
+    // this.shotSound.volume = 0.3;
+    // this.hurtSound.volume = 0.6;
+    this.backgroundSound = this.game.add.audio('background');
+    this.backgroundSound.loopFull(0.4);
   },
-
- //find objects in a Tiled layer that containt a property called "type" equal to a certain value
   findObjectsByType: function(type, map, layerName) {
     var result = new Array();
     map.objects[layerName].forEach(function(element){
       if(element.properties.type === type) {
-        //Phaser uses top left, Tiled bottom left so we have to adjust
-        //also keep in mind that some images could be of different size as the tile size
-        //so they might not be placed in the exact position as in Tiled
         element.y -= map.tileHeight;
         result.push(element);
       }
     });
     return result;
   },
-  //create a sprite from an object
   createFromTiledObject: function(element, group) {
     var sprite = group.create(element.x, element.y, element.properties.sprite);
 
@@ -89,37 +72,31 @@ Shooter.Game.prototype = {
       if (player.alive)
       {
         this.game.physics.arcade.collide(player, this.blockedLayer);
-
-        this.getOtherPlayers(player).forEach(function(player, index) {
-          console.log(player);
-          me.game.physics.arcade.overlap(me.bullets, player, me.hit, null, me);
-        });
+        me.game.physics.arcade.overlap(me.bullets, player, me.hit, null, me);
 
         var input = player.xentha.input[0];
         if(input) {
           if(input.id == 'left') {
+            player.direction = 'left';
             if(input.pressed == true) {
               player.body.velocity.x = -200;
               player.loadTexture('playerLeft');
             } else {
               player.body.velocity.x = 0;
-              player.loadTexture('player');
             }
+            player.gun.anchor.setTo(.5, -.3);
+            player.gun.scale.x = -1
           }
           if(input.id == 'right') {
+            player.direction = 'right';
             if(input.pressed == true) {
               player.loadTexture('playerRight');
               player.body.velocity.x = 200;
             } else {
               player.body.velocity.x = 0;
-              player.loadTexture('player');
             }
-          }
-          if(input.id == 'down') {
-            if(input.pressed == true) {
-              this.playerDuck(player);
-            } else {
-            }
+            player.gun.anchor.setTo(-.5, -.3);
+            player.gun.scale.x = 1
           }
           if(input.id == 'up') {
             if(input.pressed == true) {
@@ -159,7 +136,13 @@ Shooter.Game.prototype = {
     return otherPlayers;
   },
   respawn: function(player) {
-
+    player.kill();
+    var spawnPosition = this.spawnPositions[Math.floor((Math.random() * this.spawnPositions.length))];
+    player.x = spawnPosition.x;
+    player.y = spawnPosition.y;
+    player.hits = this.HITS;
+    player.loadTexture('player');
+    player.revive();
   },
   playerFire: function(player) {
     //  To avoid them being allowed to fire too fast we set a time limit
@@ -167,23 +150,19 @@ Shooter.Game.prototype = {
     {
         //  Grab the first bullet we can from the pool
         var bullet = this.bullets.getFirstExists(false);
+        bullet.checkWorldBounds = true;
 
         if (bullet && !(player.body.velocity.x == 0 && player.body.velocity.y == 0)) {
+            this.shotSound.play();
             //  And fire it
-            bullet.reset(player.x , player.y + ((player.height / 3) * 2));
+            bullet["firedBy"] = player.xentha.id;
+            bullet.reset((player.x + player.width / 2) + player.gun.anchor.x ,player.y + player.height);
 
-            if (player.body.velocity.y < 0) {
-              bullet.body.velocity.y = player.body.velocity.y - 100;
-            } else if(player.body.velocity.y > 0) {
-              bullet.body.velocity.y = player.body.velocity.y + 100;
+            if (player.direction == 'left') {
+              bullet.body.velocity.x = -400;
+            } else if(player.direction == 'right') {
+              bullet.body.velocity.x = 400;
             }
-
-            if (player.body.velocity.x < 0) {
-              bullet.body.velocity.x = player.body.velocity.x - 100;
-            } else if(player.body.velocity.x > 0) {
-              bullet.body.velocity.x = player.body.velocity.x + 100;
-            }
-            // bullet.body.velocity.x = (player.body.velocity.x < 0) ? player.body.velocity.x - 100 : player.body.velocity.X + 100;
             this.bulletTime = this.game.time.now + 200;
         }
     }
@@ -192,27 +171,36 @@ Shooter.Game.prototype = {
       //  Called if the bullet goes out of the screen
       bullet.kill();
   },
-  hit: function(player,bullet) {
+  hit: function(player, bullet) {
+    if(bullet.firedBy != player.xentha.id) {
+      this.hurtSound.play();
       var hurtVelocity = bullet.body.velocity.x;
       bullet.kill();
       me = this;
 
-      var resetVelocity = function(player) { player.body.velocity.x = 0; };
-
-      if(player.lives < 1) {
-        player.kill();
+      if(player.hits < 1) {
         player.loadTexture('playerDead');
-        // this.game.time.events.add(Phaser.Timer.SECOND * 4, respawn(player), me);
+        player.lives -= 1;
 
-        // checkIfEndGame();
+        if(player.lives < 1) {
+          player.kill();
+          this.checkIfEndGame();
+        } else {
+          this.game.time.events.add(Phaser.Timer.SECOND * 4, this.respawn, this, player);
+        }
       } else {
         player.loadTexture('playerHurt');
         // setTimeout()
         player.body.velocity.x = player.body.velocity.x + hurtVelocity;
-        // this.game.time.events.add(Phaser.Timer.SECOND * 2, resetVelocity(player), me);
-        player.lives -= 1;
+        player.hits -= 1;
+        this.game.time.events.add(Phaser.Timer.SECOND * 0.5, this.resetVelocity, this, player);
       }
+    }
 
+  },
+  resetVelocity: function(player) {
+    player.body.velocity.x = 0;
+    player.loadTexture('player');
   },
   collect: function(player, collectable) {
     //play audio
@@ -234,14 +222,19 @@ Shooter.Game.prototype = {
 
     XENTHA.playerJoined = function(data) {
       XENTHA.setLayout(XENTHA.layouts.CONTROLLER);
-      // data.players.forEach(function(xPlayer, index) {
-      var spawnPosition = me.spawnPositions[Math.floor((Math.random() * me.spawnPositions.length))];
-      var player = me.game.add.sprite(spawnPosition.x, spawnPosition.y, 'player');
+      var player = me.game.add.sprite(100, 200, 'player');
+
+      player.gun = me.game.add.sprite(0, 0, 'gun');
+      player.addChild(player.gun);
+      player.gun.anchor.setTo(-.5, -.3);
+
       me.game.physics.arcade.enable(player);
+      player.body.collideWorldBounds=true;
       player.body.gravity.y = 1000;
 
       // set default lives.
-      player.lives = this.lives;
+      player.lives = me.LIVES;
+      player.hits = me.HITS;
       player.xentha = data.player;
       me.players.push(player);
     }
@@ -296,17 +289,6 @@ Shooter.Game.prototype = {
       player.body.velocity.y -= 700;
     }
   },
-  playerDuck: function(player) {
-      //change image and update the body size for the physics engine
-      player.loadTexture('playerDuck');
-      // player.body.setSize(duckedDimensions.width, duckedDimensions.height);
-
-      //we use this to keep track whether it's ducked or not
-      player.isDucked = true;
-  },
-  render: function()
-    {
-        // this.game.debug.text(this.game.time.fps || '--', 20, 70, "#00ff00", "40px Courier");
-        // this.game.debug.bodyInfo(this.player, 0, 80);
-    }
+  render: function() {
+  }
 };
